2020_06_27 Refactoring
