+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "crm",
+  "name": "Currency calculator",
   "version": "0.1.0",
   "private": true,
   "scripts": {
